+++ apps/example-host/package.json
@@ -22,7 +22,7 @@
   },
   "dependencies": {
     "lodash": "4.16.6",
-    "zephyr-metro-plugin": "/Users/denisbelocerkovec/Developer/zephyr/zephyr-packages/libs/zephyr-metro-plugin",
+    "zephyr-metro-plugin": "../../../zephyr-packages/libs/zephyr-metro-plugin",
     "react": "19.1.0",
     "react-native": "0.80.0"
   },
