+++ apps/example-nested-mini/package.json
@@ -12,13 +12,15 @@
     "pods": "(cd ios && bundle install && (bundle exec pod install || bundle exec pod update))",
     "build:ios": "rnef bundle-mf-remote --platform ios --dev false",
     "build:android": "rnef bundle-mf-remote --platform android --dev false",
+    "build:rnc:ios": "METRO_FEDERATION_DEV=1 react-native bundle-mf-remote --platform ios --dev false",
+    "build:rnc:android": "METRO_FEDERATION_DEV=1 react-native bundle-mf-remote --platform android --dev false",
     "serve:ios": "serve dist/ios -p 8083",
     "serve:android": "serve dist/android -p 8083",
     "adbreverse": "adb reverse tcp:8083 tcp:8083"
   },
   "dependencies": {
     "lodash": "4.16.6",
-    "zephyr-metro-plugin": "/Users/denisbelocerkovec/Developer/zephyr/zephyr-packages/libs/zephyr-metro-plugin",
+    "zephyr-metro-plugin": "../../../zephyr-packages/libs/zephyr-metro-plugin",
     "react": "19.1.0",
     "react-native": "0.80.0"
   },
