+++ apps/showcase-mini/package.json
@@ -14,7 +14,7 @@
   "dependencies": {
     "lodash": "^4.17.21",
     "lottie-react-native": "^7.2.2",
-    "zephyr-metro-plugin": "/Users/denisbelocerkovec/Developer/zephyr/zephyr-packages/libs/zephyr-metro-plugin",
+    "zephyr-metro-plugin": "../../../zephyr-packages/libs/zephyr-metro-plugin",
     "react": "19.1.0",
     "react-native": "0.80.0"
   },
