+++ package.json
@@ -18,42 +18,12 @@
     "typecheck": "turbo typecheck"
   },
   "resolutions": {
-    "is-ci": "3.0.1",
-    "tslib": "2.6.2",
-    "@types/jest": "29.5.12",
-    "@types/is-ci": "3.0.0",
-    "@types/node-persist": "3.1.3",
-    "@types/proper-lockfile": "4.1.5",
-    "@types/uuid": "9.0.8",
-    "@types/debug": "4.1.12",
-    "@types/git-url-parse": "9.0.3",
-    "ts-jest": "29.1.2",
-    "@module-federation/automatic-vendor-federation": "1.0.0",
-    "zephyr-agent": "portal:/Users/denisbelocerkovec/Developer/zephyr/zephyr-packages/libs/zephyr-agent",
-    "zephyr-xpack-internal": "portal:/Users/denisbelocerkovec/Developer/zephyr/zephyr-packages/libs/zephyr-xpack-internal",
-    "zephyr-edge-contract": "portal:/Users/denisbelocerkovec/Developer/zephyr/zephyr-packages/libs/zephyr-edge-contract",
-    "buck-worker-tool": "portal:./external/metro/packages/buck-worker-tool",
-    "metro": "portal:./external/metro/packages/metro",
-    "metro-babel-register": "portal:./external/metro/packages/metro-babel-register",
-    "metro-babel-transformer": "portal:./external/metro/packages/metro-babel-transformer",
-    "metro-cache": "portal:./external/metro/packages/metro-cache",
-    "metro-cache-key": "portal:./external/metro/packages/metro-cache-key",
-    "metro-config": "portal:./external/metro/packages/metro-config",
-    "metro-core": "portal:./external/metro/packages/metro-core",
-    "metro-file-map": "portal:./external/metro/packages/metro-file-map",
-    "metro-memory-fs": "portal:./external/metro/packages/metro-memory-fs",
-    "metro-minify-terser": "portal:./external/metro/packages/metro-minify-terser",
-    "metro-resolver": "portal:./external/metro/packages/metro-resolver",
-    "metro-runtime": "portal:./external/metro/packages/metro-runtime",
-    "metro-source-map": "portal:./external/metro/packages/metro-source-map",
-    "metro-symbolicate": "portal:./external/metro/packages/metro-symbolicate",
-    "metro-transform-plugins": "portal:./external/metro/packages/metro-transform-plugins",
-    "metro-transform-worker": "portal:./external/metro/packages/metro-transform-worker",
-    "ob1": "portal:./external/metro/packages/ob1",
-    "@module-federation/runtime-core@npm:0.11.4": "patch:@module-federation/runtime-core@npm%3A0.11.4#~/.yarn/patches/@module-federation-runtime-core-npm-0.11.4-46a703d27e.patch",
-    "@react-native/community-cli-plugin@npm:0.79.0": "patch:@react-native/community-cli-plugin@npm%3A0.79.0#~/.yarn/patches/@react-native-community-cli-plugin-npm-0.79.0-d467a2c1a8.patch",
-    "react-native": "patch:react-native@npm%3A0.79.0#~/.yarn/patches/react-native-npm-0.79.0-3bd84ed11a.patch",
-    "@rnef/plugin-metro": "patch:@rnef/plugin-metro@npm%3A0.7.20#~/.yarn/patches/@rnef-plugin-metro-npm-0.7.20-f8db692ffa.patch",
+    "zephyr-agent": "portal:../zephyr-packages/libs/zephyr-agent",
+    "zephyr-xpack-internal": "portal:../zephyr-packages/libs/zephyr-xpack-internal",
+    "zephyr-edge-contract": "portal:../zephyr-packages/libs/zephyr-edge-contract",
+    "zephyr-rollx-internal": "portal:../zephyr-packages/libs/zephyr-rollx-internal"
+  },
+  "devDependencies": {
     "@biomejs/biome": "^1.9.4",
     "turbo": "^2.5.0"
   }
