+++ packages/core/src/index.ts
@@ -1,1 +1,2 @@
 export { withModuleFederation } from './plugin';
+export { updateManifest } from './plugin/manifest';
